doc: prepare for a world without whatchanged
These documents mention "whatchanged" as an example, that can be
substituted by something else.  A new hypothetical command "walken"
would come near "whatchanged" but since the latter may be going
away, we can say it would come near "version", as the fictitious
command sorts between them.  Similarly, we do not have to use
"whatchanged" as an example of a subcommand that is also implemented
in builtin/log.c file; we can instead mention "show".

Both of these changes allow us not to worry about adjusting these
places when "whatchanged" is finally removed.

Signed-off-by: Junio C Hamano <gitster@pobox.com>